test/axi: move all AXI Lite tests to separate file
